+++ src/main/frontend/src/components/layout/HeaderComponents.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Link } from "react-router-dom"
 import styled from "styled-components"
-import storage from "../../services/api/api"
+import storage from "../../services/AuthAPI/api"
 
 export default function HeaderComponents() {
 	const isLogined = storage.get({ key: "token" }) ? true : false
