+++ src/main/frontend/src/index.tsx
@@ -1,15 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import React from "react"
+import ReactDOM from "react-dom/client"
+import App from "./App"
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 
-const queryClient = new QueryClient();
-const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
-);
+const queryClient = new QueryClient()
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement)
 root.render(
-  <QueryClientProvider client={queryClient}>
-    <App />
-  </QueryClientProvider>
-);
-
+	<QueryClientProvider client={queryClient}>
+		<App />
+	</QueryClientProvider>
+)
