+++ src/main/frontend/src/router/Router.tsx
@@ -1,37 +1,30 @@
 import React from "react"
 import { BrowserRouter, Routes, Route } from "react-router-dom"
 import Header from "../pages/Header"
-import NotHeader from "../pages/NotHeader"
 import LoginPage from "../pages/LoginPage"
 import LogoutPage from "../pages/LogoutPage"
 import SignupPage from "../pages/SignupPage"
 import MainPage from "../pages/MainContainerPage"
-import RoomRegisterContext from "../hostpage/store/RoomRegisterContext"
-import HostMainPage from "../hostpage/page/HostMainPage"
-import RoomRegisterPage from "../hostpage/page/RoomRegisterPage"
-import HostUpdatePage from "../hostpage/page/HostUpdatePage"
+import RoomRegisterContext from "../store/RoomRegisterContext"
+import HostMainPage from "../pages/HostMainPage"
+import RoomRegisterPage from "../pages/RoomRegisterPage"
+import HostUpdatePage from "../pages/HostUpdatePage"
 
 const Router = () => {
 	return (
 		<BrowserRouter>
-      <RoomRegisterContext>
-		  	<Routes>
-				  <Route path="/" element={<Header />}>
-					  <Route path="MainContainerPage" element={<MainPage />} />
-					  <Route path="LoginPage" element={<LoginPage />} />
-					  <Route path="SignupPage" element={<SignupPage />} />
-				  </Route>
-				  <Route path="/" element={<NotHeader />}>
-					  <Route path="MainContainerPage" element={<MainPage />} />
-					  {/* <Route path="MainContainerPage/:hotelId" element={</>}/> */}
-					  <Route path="LogoutPage" element={<LogoutPage />} />
-					  <Route path="SignupPage" element={<SignupPage />} />
-            <Route path="/host" element={<HostMainPage/>} />
-					  <Route path="/host/update" element={<HostUpdatePage/>} />
-					  <Route path="/host/roomRegister" element={<RoomRegisterPage/>} />
-				  </Route>
-			  </Routes>
-       </RoomRegisterContext>
+			{/* <RoomRegisterContext> */}
+			<Routes>
+				<Route path="/" element={<Header />}>
+					<Route path="MainContainerPage" element={<MainPage />} />
+					<Route path="LoginPage" element={<LoginPage />} />
+					<Route path="LogoutPage" element={<LogoutPage />} />
+					<Route path="SignupPage" element={<SignupPage />} />
+					<Route path="HostMainPage" element={<HostMainPage />} />
+					<Route path="HostUpdatePage" element={<HostUpdatePage />} />
+					<Route path="RoomRegisterPage" element={<RoomRegisterPage />} />
+				</Route>
+			</Routes>
 		</BrowserRouter>
 	)
 }
